It's a little nicer now
It has some nice icons for the different options. It also has two settings: "Always On Top" will make it stay on top of other apps and also make it semi-transparent. "Change Font Size" will change the clock's font to match whatever is in the Numeric Up Down beside it.